requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@4429b564fd8f031deaccd60caa5be7320e248816
discord-ext-slash==0.5.0
discord.py==1.6.0
dnspython==2.1.0
idna==3.1
multidict==5.1.0
numexpr==2.7.3
numpy==1.20.2
pycparser==2.20
pymongo==3.11.3
PyNaCl==1.4.0
python-dotenv==0.16.0
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
